gitignore, fix bugs
- Изменен gitignore
- Убрана конструкция try catch с сервисов
- Restful API
- Поправил тесты
