--- Sol.Api/Services/IStudentService.cs
@@ -1,12 +1,0 @@
-﻿using Sol.Domain.Dto;
-
-namespace Sol.Api.Services;
-
-public interface IStudentService
-{
-    public Task<bool> CreateUser(StudentDto student);
-    
-    public Task<bool> EditUser(StudentDto student);
-    
-    public Task<bool> DeleteUser(int id);
-}
